Fix critical bug preventing decimal frequencies in edit service endpoint
- Changed frequency_per_year parameter type from int to float in edit_service endpoint
- This was causing decimal values like 1.5 to be truncated to 1 when editing services
- The add_service endpoint was already correctly typed as float
- Now decimal frequencies work properly for both adding and editing services

+++ web_app.py
@@ -282,7 +282,7 @@ async def edit_service(
     service_index: int = Form(...),
     service_name: str = Form(...),
     unit_cost: float = Form(...),
-    frequency_per_year: int = Form(...),
+    frequency_per_year: float = Form(...),
     inflation_rate: float = Form(...),
     service_type: str = Form(...),  # "recurring" or "discrete"
     start_year: Optional[int] = Form(None),
